Fixing a few problems with the DIII-D machine mapping
magnetics_hardware requires a pulse to be specified

CER conversion was never finished.  Several measurements are now
being filled, but others will require more work (place holders added).

The gas_injection IDS was updated to move the position of valves.  The
mapping was adjusted to account for this.

The d3d.json file had several entries that requested a specific time
slice, which does not fit with the use of everything else so they have
been removed.