Fix DIII-D machine mapping for F-coil currents
IMAS specifies the current should be the measured value for a single
turn, but MDS+ stores A-turns.  This only impacts F-coils and is
related to a quirk for how the coils are specified in EFIT.

Also fixed a bad IDS entry and reordered coils in pf_active to match
the OMFIT ordering.

This MR should be accompanied by corresponding fixes in OMFIT so that
k-files are still formed properly (from the OMAS machine mappings).