Update to use IMAS 3.40.0
The primary changes (that I am aware of) in this version of IMAS are:
1) Renamed gyrokinetic ids to gyrokinetic_local
2) Restructured pulse_schedule ids to eliminate
pulse_schedule.*.*.data and pulse_schedule.*.*.time entries
3) Extended the equilibrium ids to include more coordinate options
for relevant ids (R, Z, psi, phi, rho_tor_norm) and added an overall
chi_squared_reduced metric

Some EFIT machine mapping fixes are also included in this PR along
with a name change in some mhdin.dat file variables.

test_omas_dynamic_machine was turned off in this PR because of an
open issue with omas that results in crashing even on older versions
of the master branch.

This PR will be accompanied by an OMFIT PR as well.