Update to use IMAS 3.40.1
This seems like a pretty small change overall.  I haven't seen much
change in the DD and all tests are still passing.

git seems more confused about the update than should be necessary,
but the changes were all automated to I'm planning to leave that be.

The OMFIT regression tests need to be checked before this is merged.