Adding several mhdin versions for KSTAR
These were all squashed together in a file they shared which wasn't
very clear or well organized. I've tried to order them based on what
little history could be gleaned from the file, but it's just a best
guess. None line up perfectly with the dprobe.dat that was already
part of this repo, so they should still be used with some caution.