Update tests
The IO seems to be using different branch conventions in the open source repo
so "develop/3" no longer exists and "develop" doesn't appear to be
a complete copy of the source (or has a bug). This is being replaced
by "main" which seems like the most stable. "release/3" could be
a better replacement, but the distinction isn't entirely clear.