Integrate magnetics_weights mapping into magnetics_*_data
This removes the standalone mapping function that put the weights
into the equilibrium constraints and instead uses them to set the
validity and uncertainty of the diagnostic data. This is more
appropriate since these weights are static and _efit.json already
includes a mapping for the equilibrium constraint weights.

Also changing how the option to zero the reference flux loop works.
This is now disabled by default and when activated shifts all of the
probes to be relative flux.