Ensure that flux loop data is consistent with IMAS
DIII-D measures and stores all other flux loops differentially
relative to PSF1A and has since at least 1988. Unfortunately the
only differential flux loops supported in IMAS are diamagnetic so
the poloidal loops must be converted to total flux in order to be
valid. This is now the default behavior, but an option is available
to skip the conversion if IMAS adherence isn't a concern. The option
to zero the reference flux loop has been completely removed since it
throws out important information. Even if IMAS adds support for
differential poloidal flux loops, it probably won't be in a version
supported by OMAS in the foreseeable future.

Also moving the flux loop and probe compensations after the
uncertainties are feteched to avoid attempting their application
for channels that are disabled.